DRAW-348 refactor: Domain에 있는 객체 Core로 임시 이관
