+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/event/BroadcastEvent.kt
@@ -1,7 +1,7 @@
 package com.xorker.draw.websocket.broker.event
 
 import com.xorker.draw.room.RoomId
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.SessionMessage
 
 data class BroadcastEvent(
     val roomId: RoomId,
