+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/event/UnicastEvent.kt
@@ -1,7 +1,7 @@
 package com.xorker.draw.websocket.broker.event
 
 import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.SessionMessage
 
 data class UnicastEvent(
     val userId: UserId,
