+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/game/MafiaGamePlayerListMessage.kt
@@ -2,8 +2,8 @@ package com.xorker.draw.websocket.message.response.dto.game
 
 import com.xorker.draw.mafia.MafiaPlayer
 import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.ResponseAction
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.ResponseAction
+import com.xorker.draw.websocket.message.response.SessionMessage
 
 data class MafiaGamePlayerListMessage(
     override val body: MafiaGamePlayerListBody,
