+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/game/MafiaGameTurnInfoMessage.kt
@@ -1,8 +1,8 @@
 package com.xorker.draw.websocket.message.response.dto.game
 
 import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.ResponseAction
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.ResponseAction
+import com.xorker.draw.websocket.message.response.SessionMessage
 import java.time.LocalDateTime
 
 data class MafiaGameTurnInfoMessage(
