+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/game/MafiaGameVoteStatusMessage.kt
@@ -1,8 +1,8 @@
 package com.xorker.draw.websocket.message.response.dto.game
 
 import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.ResponseAction
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.ResponseAction
+import com.xorker.draw.websocket.message.response.SessionMessage
 
 data class MafiaGameVoteStatusMessage(
     override val body: MafiaGameVoteStatusBody,
