+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/phase/MafiaPhaseEndMessage.kt
@@ -1,7 +1,7 @@
 package com.xorker.draw.websocket.message.response.dto.phase
 
-import com.xorker.draw.websocket.ResponseAction
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.ResponseAction
+import com.xorker.draw.websocket.message.response.SessionMessage
 import java.time.LocalDateTime
 
 data class MafiaPhaseEndMessage(
