+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/phase/MafiaPhaseWaitMessage.kt
@@ -1,8 +1,8 @@
 package com.xorker.draw.websocket.message.response.dto.phase
 
 import com.xorker.draw.room.RoomId
-import com.xorker.draw.websocket.ResponseAction
-import com.xorker.draw.websocket.SessionMessage
+import com.xorker.draw.websocket.message.response.ResponseAction
+import com.xorker.draw.websocket.message.response.SessionMessage
 import com.xorker.draw.websocket.message.response.dto.game.MafiaPlayerResponse
 
 data class MafiaPhaseWaitMessage(
