+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/parser/WebSocketResponseParser.kt
@@ -1,8 +1,8 @@
 package com.xorker.draw.websocket.parser
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.xorker.draw.websocket.SessionMessage
 import com.xorker.draw.websocket.exception.ExceptionMessage
+import com.xorker.draw.websocket.message.response.SessionMessage
 import org.springframework.stereotype.Component
 
 @Component
